refactorizaton of the transform/quant path.
This patch involves scaling down the DCT/IDCT coefficients from int32_t to
int16_t as they can be accommodated on int16_t without any introduction of
encode error, this allows us to clean up lots of DCT/IDCT intermediate
buffers, optimize enode efficiency for different cli options including noise
reduction by reducing data movement operations, accommodating more number of
coefficients in a single register for SIMD operations. This patch include all
necessary changes for the transfor/quant path including unit test code.

+++ source/common/quant.h
@@ -83,8 +83,8 @@ protected:
 
     bool               m_useRDOQ;
     int64_t            m_psyRdoqScale;
-    int32_t*           m_resiDctCoeff;
-    int32_t*           m_fencDctCoeff;
+    int16_t*           m_resiDctCoeff;
+    int16_t*           m_fencDctCoeff;
     int16_t*           m_fencShortBuf;
 
     enum { IEP_RATE = 32768 }; /* FIX15 cost of an equal probable bit */
